moved InputSelect to separate component, moved styles to separate style file
