+++ src/react/src/components/ClaimFacilityAdditionalData.jsx
@@ -4,20 +4,8 @@ import { connect } from 'react-redux';
 import Typography from '@material-ui/core/Typography';
 import TextField from '@material-ui/core/TextField';
 import InputLabel from '@material-ui/core/InputLabel';
-import Switch from '@material-ui/core/Switch';
-import noop from 'lodash/noop';
-import find from 'lodash/find';
 import flow from 'lodash/flow';
-import stubFalse from 'lodash/stubFalse';
 import map from 'lodash/map';
-import filter from 'lodash/filter';
-import includes from 'lodash/includes';
-import isNull from 'lodash/isNull';
-import Select from 'react-select';
-import Creatable from 'react-select/creatable';
-
-import ShowOnly from './ShowOnly';
-import CreatableInputOnly from './CreatableInputOnly';
 
 import {
     updateClaimASector,
@@ -37,27 +25,7 @@ import {
 } from '../util/styles';
 
 import { claimAFacilityAdditionalDataFormFields } from '../util/constants';
-
-import COLOURS from '../util/COLOURS';
-
-const infoTitleStyle = Object.freeze({
-    paddingBottom: '10px',
-    color: COLOURS.NEAR_BLACK,
-    fontWeight: 'bold',
-});
-
-const selectStyles = Object.freeze({
-    input: provided =>
-        Object.freeze({
-            ...provided,
-            padding: '10px',
-        }),
-    menu: provided =>
-        Object.freeze({
-            ...provided,
-            zIndex: '2',
-        }),
-});
+import InputSection from '../components/InputSection';
 
 const {
     sectorsForm,
@@ -67,131 +35,6 @@ const {
     localLanguageNameForm,
     localLanguageNameDesc,
 } = claimAFacilityAdditionalDataFormFields;
-
-const InputSection = ({
-    value,
-    multiline,
-    onChange,
-    hasSwitch = false,
-    switchValue = null,
-    onSwitchChange = noop,
-    disabled = false,
-    isSelect = false,
-    isMultiSelect = false,
-    isCreatable = false,
-    selectOptions = null,
-    hasValidationErrorFn = stubFalse,
-    aside = null,
-    selectPlaceholder = 'Select...',
-}) => {
-    let SelectComponent = null;
-
-    const asideNode = (
-        <ShowOnly when={!isNull(aside)}>
-            <aside style={claimedFacilitiesDetailsStyles.asideStyles}>
-                {aside}
-            </aside>
-        </ShowOnly>
-    );
-
-    if (isSelect) {
-        const selectValue = (() => {
-            if (!isCreatable && !isMultiSelect) {
-                return find(selectOptions, ['value', value]);
-            }
-
-            if (!isCreatable && isMultiSelect) {
-                return filter(selectOptions, ({ value: option }) =>
-                    includes(value, option),
-                );
-            }
-
-            if (isCreatable && isMultiSelect) {
-                return map(value, s => ({ value: s, label: s }));
-            }
-
-            // isCreatable && !isMultiSelect creates an option object from the value
-            // if it doesn't exist in the options list
-            const option = find(selectOptions, ['value', value]);
-            return (
-                option || {
-                    value,
-                    label: value,
-                }
-            );
-        })();
-
-        if (isCreatable) {
-            SelectComponent = selectOptions ? Creatable : CreatableInputOnly;
-        } else {
-            SelectComponent = Select;
-        }
-
-        return (
-            <div
-                style={
-                    claimedFacilitiesDetailsStyles.inputSectionStylesWithPadding
-                }
-            >
-                {asideNode}
-                <SelectComponent
-                    onChange={onChange}
-                    value={selectValue}
-                    options={selectOptions}
-                    disabled={disabled}
-                    styles={selectStyles}
-                    isMulti={isMultiSelect}
-                    placeholder={selectPlaceholder}
-                />
-            </div>
-        );
-    }
-
-    return (
-        <div style={claimedFacilitiesDetailsStyles.inputSectionStyles}>
-            <InputLabel
-                style={claimedFacilitiesDetailsStyles.inputSectionLabelStyles}
-            >
-                {hasSwitch ? (
-                    <span
-                        style={
-                            claimedFacilitiesDetailsStyles.switchSectionStyles
-                        }
-                    >
-                        <Switch
-                            color="primary"
-                            onChange={onSwitchChange}
-                            checked={switchValue}
-                            style={{ zIndex: 1 }}
-                        />
-                        Publicly visible
-                    </span>
-                ) : null}
-            </InputLabel>
-            {asideNode}
-            <TextField
-                variant="outlined"
-                style={claimedFacilitiesDetailsStyles.inputSectionFieldStyles}
-                value={value}
-                multiline={multiline}
-                rows={6}
-                onChange={onChange}
-                disabled={disabled}
-                error={hasValidationErrorFn()}
-            />
-        </div>
-    );
-};
-
-// InputSection.propTypes = {
-//     classes: shape({
-//         switchSectionStyles: string.isRequired,
-//         inputSectionFieldStyles: string.isRequired,
-//         inputSectionLabelStyles: string.isRequired,
-//         inputSectionStyles: string.isRequired,
-//         asideStyles: string.isRequired,
-//     }).isRequired,
-// };
 
 function ClaimFacilityAdditionalData({
     sectors,
@@ -214,7 +57,10 @@ function ClaimFacilityAdditionalData({
         <>
             <div style={claimAFacilitySupportDocsFormStyles.inputGroupStyles}>
                 <InputLabel htmlFor={sectorsForm.id}>
-                    <Typography variant="display1" style={infoTitleStyle}>
+                    <Typography
+                        variant="display1"
+                        style={claimedFacilitiesDetailsStyles.infoTitleStyle}
+                    >
                         {sectorsForm.label}
                     </Typography>
                 </InputLabel>
@@ -229,11 +75,16 @@ function ClaimFacilityAdditionalData({
                     isMultiSelect
                     selectOptions={sectorOptions || []}
                     selectPlaceholder={sectorsForm.placeholder}
+                    classes={claimedFacilitiesDetailsStyles}
+                    isClaimFacilityAdditionalDataPage
                 />
             </div>
             <div style={claimAFacilitySupportDocsFormStyles.inputGroupStyles}>
                 <InputLabel htmlFor={numberOfWorkersForm.id}>
-                    <Typography variant="display1" style={infoTitleStyle}>
+                    <Typography
+                        variant="display1"
+                        style={claimedFacilitiesDetailsStyles.infoTitleStyle}
+                    >
                         {numberOfWorkersForm.label}
                     </Typography>
                 </InputLabel>
@@ -253,7 +104,10 @@ function ClaimFacilityAdditionalData({
             </div>
             <div style={claimAFacilitySupportDocsFormStyles.inputGroupStyles}>
                 <InputLabel htmlFor={localLanguageNameForm.id}>
-                    <Typography variant="display1" style={infoTitleStyle}>
+                    <Typography
+                        variant="display1"
+                        style={claimedFacilitiesDetailsStyles.infoTitleStyle}
+                    >
                         {localLanguageNameForm.label}
                     </Typography>
                 </InputLabel>
